delete css and file renaming
