+++ .gitignore
@@ -21,6 +21,7 @@ yarn-error.log*
 /src/forms/MultiChoiceButton.css
 /src/mcode-pilot/TreatmentOptionsSelector/TreatmentOptionsSelector.css
 /src/mcode-pilot/TreatmentOptionsVisualizer/TreatmentOptionsVisualizer.css
+/src/mcode-pilot/**/*.css
 /src/notes/FluxNotesEditor.css
 /src/notes/NoteAssistant.css
 /src/notes/fillFieldComponents/MenuItemSetFillForPlaceholder.css
