--- src/mcode-pilot/TreatmentOptionsVisualizer/TreatmentOptionsVisualizer.css
@@ -1,3 +1,0 @@
-div.treatment-options-visualizer__info {
-  font-size: 0.8em;
-  margin: 5px 25px; }
